content/html/content/test/test_bug555840.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=555840
-->
<head>
  <title>Test for Bug 555840</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=555840">Mozilla Bug 555840</a>
<p id="display"></p>
<div id="content" style="display: none">
  <datalist>
  </datalist>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 555840 **/

function checkClassesAndAttributes()
{
  var d = document.getElementsByTagName('datalist');
  is(d.length, 1, "One datalist has been found");

  d = d[0];
  ok(d instanceof HTMLDataListElement,
     "The datalist should be instance of HTMLDataListElement");

  ok('options' in d, "datalist has an options IDL attribute");

  ok(d.options, "options IDL attribute is not null");
  ok(!d.getAttribute('options'), "datalist has no options content attribute");

  ok(d.options instanceof HTMLCollection,
     "options IDL attribute should be instance of HTMLCollection");
}

function checkOptions()
{
  var testData = [
    /* [ Child list, Function modifying children, Recognized options ] */
    [['option'], null, 1],
    [['option', 'option', 'option', 'option'], null, 4],
    /* Disabled options are not valid. */
    [['option'], function(d) { d.childNodes[0].disabled = true; }, 0],
    [['option', 'option'], function(d) { d.childNodes[0].disabled = true; }, 1],
    /* Non-option elements are not recognized. */
    [['input'], null, 0],
    [['input', 'option'], null, 1],
    [['input', 'textarea'], null, 0],
    /* .value and .label are not needed to be valid options. */
    [['option', 'option'], function(d) { d.childNodes[0].value = 'value'; }, 2],
    [['option', 'option'], function(d) { d.childNodes[0].label = 'label'; }, 2],
    [['option', 'option'], function(d) { d.childNodes[0].value = 'value'; d.childNodes[0].label = 'label'; }, 2],
  ];

  var d = document.getElementsByTagName('datalist')[0];

  for each (data in testData) {
    for each (e in data[0]) {
      d.appendChild(document.createElement(e));
    }

    /* Modify children. */
    if (data[1]) {
      data[1](d);
    }

    is(d.options.length, data[2],
       "The number of recognized options should be " + data[2])

    /* Cleaning-up. */
    for (; d.firstChild; d.removeChild(d.firstChild));
  }
}

checkClassesAndAttributes();
checkOptions();

</script>
</pre>
</body>
</html>
